perf(socks): 优化 Fake IP 机制，将 Hash 从 BigInteger 改为 long
- SocksRpcContract: 将 fakeDict 的 Key 和 fakeEndpoint 的参数从 BigInteger 改为 long，显著提升性能并减少内存开销；新增 fakeHost 与 parseFakeHostHash 辅助方法。
- SocksTcpUpstream: 实现高效的 64 位混合哈希算法 (mix hash)，用于生成终结点的混淆 Hash。
- Socks5CommandRequestHandler: 适配新的 long 类型 Hash 恢复原始终结点。
- start.sh: 调整 outlandServers 配置，移除 1.1.1.1 依赖。
- 测试: 更新集成测试以匹配新的接口定义。

+++ rxlib/src/main/java/org/rx/util/rss/RssClient.java
@@ -52,7 +52,6 @@ import org.rx.util.function.QuadraFunc;
 import org.rx.util.function.TripleAction;
 
 import java.io.OutputStream;
-import java.math.BigInteger;
 import java.net.HttpURLConnection;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
@@ -134,7 +133,7 @@ public final class RssClient {
         }
 
         @Override
-        public void fakeEndpoint(BigInteger hash, String realEndpoint) {
+        public void fakeEndpoint(long hash, String realEndpoint) {
             delegate.fakeEndpoint(hash, realEndpoint);
         }
